refactor(share) optimize permissions endpoint to make it faster

+++ apps/Tableau/pages/api/permissions/share.ts
@@ -7,6 +7,8 @@ import {
     editShareablePermissionCb,
     withAuth,
 } from 'shared-libs'
+import { mockSession } from 'next-auth/client/__tests__/helpers/mocks'
+import user = mockSession.user
 
 interface IRequestObject {
     boardId: string
@@ -29,21 +31,60 @@ export default async function handler(
 
         const ShareableRolesPermissions = new Map()
 
+        const userBoardSharingDatabase = await prisma.boardUserSharing.findMany(
+            {
+                include: { user: true, board: { include: { user: true } } },
+            }
+        )
+
+        const currentUser = await prisma.user.findFirst({
+            where: { email: session.user!.email },
+        })
+
+        if (!currentUser) return res.status(400).send('Bad Request')
+
         for (let idx = 0; idx < userBoardSharing.length; idx++) {
             const { id } = userBoardSharing[idx]
+            const userBoardSharingDatabaseToEdit =
+                userBoardSharingDatabase.find(
+                    (userBoardSharingSingle) => userBoardSharingSingle.id === id
+                )
+
+            if (!userBoardSharingDatabaseToEdit) return
+
+            const userBoardSharingDatabaseOfCurrentUser =
+                userBoardSharingDatabase.find(
+                    (userBoardSharingSingle) =>
+                        userBoardSharingSingle.boardId ===
+                            userBoardSharingDatabaseToEdit.boardId &&
+                        userBoardSharingSingle.userId === currentUser.id
+                )
+
             const edit = await editShareablePermissionCb({
                 req,
                 res,
                 authOptions,
                 prisma,
-                params: { id },
+                params: {
+                    id,
+                    currentUser: currentUser!,
+                    boardUserSharingToEdit: userBoardSharingDatabaseToEdit,
+                    boardUserSharingOfCurrentUser:
+                        userBoardSharingDatabaseOfCurrentUser,
+                },
             })
             const remove = await deleteShareablePermissionCb({
                 req,
                 res,
                 authOptions,
                 prisma,
-                params: { id },
+                params: {
+                    id,
+                    currentUser: currentUser!,
+                    boardUserSharingToEdit: userBoardSharingDatabaseToEdit,
+                    boardUserSharingOfCurrentUser:
+                        userBoardSharingDatabaseOfCurrentUser,
+                },
             })
 
             ShareableRolesPermissions.set(id, {
